# docs/handoff/ups-expresosapp-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/vehicles": {
      "post": {
        "operationId": "VehiclesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Vehicle plate or code already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new vehicle",
        "tags": [
          "Admin Vehicles"
        ]
      },
      "get": {
        "operationId": "VehiclesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "description": "Page number (1-based)",
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "description": "Items per page",
            "required": false,
            "name": "limit",
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of vehicles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiclePaginatedResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all vehicles with pagination",
        "tags": [
          "Admin Vehicles"
        ]
      }
    },
    "/admin/vehicles/{id}": {
      "get": {
        "operationId": "VehiclesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Vehicle not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get vehicle details",
        "tags": [
          "Admin Vehicles"
        ]
      },
      "patch": {
        "operationId": "VehiclesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Vehicle not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a vehicle",
        "tags": [
          "Admin Vehicles"
        ]
      }
    },
...
  },
  "info": {
    "title": "UPS ExpresosApp API",
    "description": "Backend API for UPS ExpresosApp MVP - Institutional transport management system for Universidad Politecnica Salesiana.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string",
            "description": "Vehicle license plate",
            "example": "ABC-1234"
          },
          "code": {
            "type": "string",
            "description": "Internal vehicle code",
            "example": "V001"
          },
          "capacity": {
            "type": "number",
            "description": "Passenger capacity",
            "minimum": 1,
            "example": 40
          },
          "status": {
            "type": "string",
            "description": "Vehicle status",
            "enum": [
              "ACTIVE",
              "MAINTENANCE",
              "INACTIVE"
            ],
            "default": "ACTIVE"
          }
        },
        "required": [
          "plate",
          "code",
          "capacity"
        ]
      },
      "VehicleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "plate": {
            "type": "string",
            "example": "ABC-1234"
          },
          "code": {
            "type": "string",
            "example": "V001"
          },
          "capacity": {
            "type": "number",
            "example": 40
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "MAINTENANCE",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "plate",
          "code",
          "capacity",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "VehiclePaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleResponseDto"
            }
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string",
            "description": "Vehicle license plate",
            "example": "ABC-1234"
          },
          "code": {
            "type": "string",
            "description": "Internal vehicle code",
            "example": "V001"
          },
          "capacity": {
            "type": "number",
            "description": "Passenger capacity",
            "minimum": 1,
            "example": 40
          },
          "status": {
            "type": "string",
            "description": "Vehicle status",
            "enum": [
              "ACTIVE",
              "MAINTENANCE",
              "INACTIVE"
            ],
            "default": "ACTIVE"
          }
        }
      },
...
    }
  }
}